GhostPDF - handle null objects from DecodeParams array
Bug #705626 "Ghostscript 9.56 not renders PDF if PDF Extraction Permissions denies page extraction"

The PDF file has a number of images compressed with the JBIG2 filter
and with a DecodeParams array of [null]. Which is, of course, pointless.

However we were throwing an error because the DecodeParams entry was not
a dictionary type, instead we should simply ignore it, which this commit
implements.