oss-fuzz 48305: Use ref counting for subclass child device
The .currentoutputdevice (custom) operator leaves the chance that PS VM still
has a reference to the "child" of a subclassed device. In normal operation,
this isn't a problem, but on an error, that PS reference may persist until
after the unsubclass operation, meaning the "child" is freed while there is
still an active reference to it.

Change the code to use reference counting to ensure that works. This means
considerable fiddling to ensure that, after subclassing, anything sent to the
(now moribund) child device doesn't end up affecting the parent.