CUPS/PWG/Apple Raster output device: Support for PPD-less use
PPD (PostScript Printer Description) files are deprecated for many
years now, after having PDF-based printing for more than a decade and
will finally get abolished completely with CUPS 3.x due end of
2023. Instead, printing will get completely IPP-based.

cups-filters is currently prepared for the transition to this New
Architecture of printing and this requires some changes in the
Ghostscript output device driver (cups/gdevcups.c) for accesing all
parameters without the need of a PPD file. These changes are done with
this commit.

For this the following new command line options got added:

CUPS Raster version (compression)
---------------------------------

The integer option "-dcupsRasterVersion=X" determines whether CUPS
Raster output is compressed (X = 2) or uncompressed (X = 3, the
default). This only applies to CUPS Raster ("-sDEVICE=cups",
MediaClass != ¨PwgRaster") and not to PWG Raster or Apple Raster/URF.

This option is equivalent to the "*cupsRasterVersion" PPD file
keyword.

Back side orientation for duplex printing
-----------------------------------------

The string option "-scupsBackSideOrientation=X" with X being one of
"Normal" (the default), "ManualTumble", "Rotated", or "Flipped" tells
the orientation of the back side raster images when printing
double-sided. Many printers feed the paper in another direction for
printing the back side and so for example start the printing at the
bottom of the page and not at the top, to make the duplexing mechanics
simpler, smaller, and cheaper. See

    https://openprinting.github.io/cups/doc/spec-ppd.html#cupsBackSide

for the meaning of the orientations. This option is equivalent to the
"*cupsBackSide" PPD file keyword.

Flipping top/bottom margins on back sides when printing duplex
--------------------------------------------------------------

The boolean option "-dcupsBackSideFlipMargins" tells that when
printing double-sided for the back sides the top and the bottom
margins need to get swapped. See

    https://openprinting.github.io/cups/doc/spec-ppd.html#APDuplexRequiresFlippedMargin

for more information. This option is equivalent to the
"*APDuplexRequiresFlippedMargin" PPD file keyword.

Manual copies needed
--------------------

The boolean option "-dcupsManualCopies" makes the raster image be sent
repeatedly for printers which cannot generate copies internally.

This option is equivalent to the "*cupsManualCopies" PPD file keyword.

Note that if you supply both a PPD file (via "PPD" environment
variable) and use these new command line options, the values defined
in the PPD file have priority.