GhostPDF - fix memory leak with OSS-fuzz #48462
Discovered by Robin while working on the OSS-fuzz bug. This is due to
the change to objects not always being pointers to structures any more.
When dereferencing an object in a dictionary we replace the indirect
object with the returned dereferenced object. However in the case of an
error dereferencing the object we get a null object returned.

Before replacing the object in the dictionary we need to specifically
test to see it is a pointer to a structure, because we check the object
number in the structure. The problem was caused by the fact that the
code for dealing with a non-pointer object didn't count down the object
in the dictionary before replacing it, leaving it dangling.