Device subclassing with transparency compositor
The commit for optionally keeping subclassed devices 'ahead' of the
transparency compositor f20fac9d4d505be579ca4c5872233b3ec2d85f85
missed the possibility that the device ahead of the compositor may
itself have parent devices.

This is relevant because the transparency compositor can change colour
model, and that needs to be reflected in the 'first' device in the
chain. So if the device has one or more parent devices we need to update
those as well.

The problem was demonstrated by testing with the compile-time option
FORCE_TESTING_SUBCLASSING in base/gdevkrnlsclass.c defined so that we
have more subclassing devices enabled.