OSS-fuzz bug #48527
Still fallout from the 'objects aren't always a pointer' commit. In this
case the loop should stop based on the type as well as the object
number. If the Parent isn't a dictionary (it should always be) then stop
checking backwards.

The object is in fact a null, because the indirect reference is to a
non-existent object.