OSS_fuzz #48564
Yet more fallout from the change to objects not always being pointers
to structures. using object->type doesn't work if object isn't a
pointer to a struct, we need to use pdfi_type_of instead.

Since this keeps happening, audit all occurrences of ->type and .type in
the pdfi code and make sure we use the pdfi_type_of() function where
needed. A couple of these are in code not compiled in due to #if 0
pre-processor directives, but fix them anyway.