oss-fuzz 48547/48528/48526: Various TTF hinter fixes
A couple of opcode functions in the bytecode interpreter were not bounds
checking the values they used.

If a bytcode execution context initialisation encountered an out of memory
error part way through, the remainder of the context would be left
uninitialised which could then lead to a later crash when cleaning up the
partially initialised context.