oss-fuzz 48437: Add bounds check on size of loca table data
The test case has a TTF that the loca size has been set to an enormous value.
While the loca size is a 32 bit unsigned number (so the value in question is
"valid") it should follow a specific relationship with the number of glyphs
that the font contains, and that is a an unsigned 16 bit value.

So, we can sanity check the size of the loca table based on that condition.

This may or may not solve the oss-fuzz timeout, but it does improve things
considerably.