GhostPDF - add minimal support for /F in streams
Bug #705627 "New PDF parser ignores /F in streams"

The /F key in a stream dictionary indicates that the actual stream is to
be found i an external file. This seems to me to be of limited utility;
it breaks portability because the files have to be in specific locations
or they cannot be located. In Ghostscript the interpreter must also be
instructed to permit opening of these files.

Nevertheless, as the report states, this did work (up to a point) with
the old interpreter so this commit adds the same degree of support. If
the /F key is associated with a file specification string value then
we will attempt to open that file and us it in place of the embedded
stream.

In addition, if the /F value is a file specification dictionary then we
will attempt to read a file specification string (also /F) from that
dictionary and, if successful, use that. The old interpreter does not
appear to have supported file specification dictionaries here at all.