Store /EPSBoundingBoxString in global VM
When doing the special processing for EPS files, we store the string from
the %%BoundingBox comment in systemdict so it can be accessed later for
doing fit page functionality.

Problem is, the special EPS handling is done in between save/restore operations,
meaning the string is restored away while a reference to it remains in
systemdict, potentially causing problems.

Moving the string to global VM resolves the problem