GhostPDF - handle page numbers in /Dest
No reproducer for this. An Outline with a Dest which involved a page
number, such as "/Dest [ 0/XYZ 68 743 0]" didn't work because the
code to handle page view was expecting a page dictionary, not a page
number.

According to the PDF specification (p582 of the PDF 1.7 version) the
first element is an indirect reference to a page object. Clearly
this isn't the case here. Yet Acrobat happily uses the outline Dest
(the file has only one page, so its unclear what happens in a file
with multiple pages)

Page numbers are actually much easier to handle, so I've chosen to
add the required code here to deal with page numbers as well as
page dictionaries. We may need to revisit this in future.

Noticed in passing that we were accidentally swallowing error codes
in a return, so fix that too. This was causing incorrect Outlines to
be sent when an error occurred.