GhostPDF improve hashing of pattern IDs
Simply adding the number of components to the object number doesn't work
well, because (eg) object number 19 with 3 components is then exactly
the same as object number 21 with 1 component.

Applying the hash function to the data works much better.

In the long run we may need something better than the simple hash in
order to prevent hash collisions like this.