FAPI - Don't free memory on error if it doesn't belong to FAPI
OSS-fuzz #44994 with this configuration:

-K1048576 -r200x200 -sBandListStorage=memory -dMaxBitmap=0 -dBufferSpace=450k -dMediaPosition=1 -dcupsColorSpace=1 -dSAFER -dNOPAUSE -dBATCH -dNOINTERPOLATE -dNOMEDIAATTRS -sOutputFile=/temp/out.tif -sDEVICE=tiff24nc

and SHA e69805c35afa4544ab4598f173c802349d86be6d caused a seg fault when
freeing the font's sfnt.data in pdfi_free_font_truetype().

This is because the memory had already been freed. The memory was freed
in gs_fapi_ft_get_scaled_font() due to an error calling new_inc_int() at
line 1315. The error would free 'own_font_data' which was retrieved
from the font at line 1281, and returned memory which was allocated
by the pdfi interpreter, not by FAPI.

The code round line 1282 sets 'data_owned' to false, and other places
which detect errors do not free the memory if data_owned is false so
this looks like a simple oversight.