OSS-fuzz #4090 - limit angle operands to arc
The file executes an arc operation with angle2=1 and angle1=3.5x10^21.
The spec says that if angle2 is less than angle1 we should increase it
by multiples of 360 until it is greater than or equal to angle1.

The problem is that the accuracy of the double representation is not
sufficient to detect the addition of 360, so we go round the loop
adding 360 to angle2 trying to get it larger than angle1 forever.

Trying to detect the point at which this would occur is hard, and
probably error-prone across platforms, so we choose to limit the two
angles to the platform maximum integer, less 360 to allow a full
circle between the angles.

If either angle exceeds that we will return a limitcheck error.