GhostPDF - free dangling path on exit
Since we now store the current path, because in PDF it is not part of
the graphics state and not subject to gsave/grestore, it is possible to
reach the end of job with unused path segments which are not freed when
the graphics state is.

So free any dangling path on exit.