GhostPDF - fix cleanup when dash array isn't an array
Probably crept in when ensuring we take a proper reference to objects
instead of relying on them remaining on the stack. If we got a setdash
where the array operator wasn't an array we tried to pop the array from
the stack as part of the error handling. Since we've already removed it
from the stack this is incorrect, we should count it down instead.