Bug 705652: Add user defined fontmap to permit read paths
The Postscript interpreter parses the Fontmap file during initialisation, so
before file access permission enforcement has been activated, so a user
defined fontmap file need not be added to the permit-file-read list.

But pdfi parses the fontmap on demand, thus after file access permissions are
active. So when calling pdfi from Postscript, the Postscript interpreter
needs to add the file and path to the permit-file-read list so pdfi can
access it.