OSS-fuzz #49174 - memory leak in tiffsep/tiffsep1
The initial problem is that the tiffsep devices, for a single output
file, do not close and cleanup at the end of each page, that is deferred
until the device is closed. (File-per-page output does clean up after
every page and doesn't suffer from at least the reported leak.)

But in order to properly close and cleanup, we need to know how many
spot colours there were. That is retrieved from the DeviceN params, but
that simply doesn't work, because the DeviceN params are reset at the
end of each page, so by the time the device is closed we think there
are no spots.

This commit adds a new parameter to the tiffsep devices which simply
records the number of spot inks at the end of the page, for proper
cleanup. If we have a multiple page input, and the number of inks are
not the same on each page, then the code already aborts with an error so
we can be assured this number is correct.

That allows cleanup of the dangling file, but there were more leaks.

The 'private' data we attach to each TIFF output file was never freed,
because the TIFF library expects the client to do that itself, and we
never did.

To address these I've added a couple of new functions to our interface
to the TIFF library. One allows us to retrieve the filename for a given
output file, to avoid us having to generate it all over again. To do
that we would have needed to preserve all the DeviceN data and I really
didn't want to go down that road. Since the only reason we needed it was
to rebuild the filename, and the name is already in the TIFF structure,
and there's an API call to retrieve it, simply use that.

The second function just frees the private memory we allocated in
tiff_from_filep(). Note that we are using gs_free() here, not
gs_free_object() because the allocation uses gs_malloc(). If that is
changed we'll need to change the free as well.

Because tiffsep and tiffsep1 share much common code, the changes to the
tiffsep1 device needed to be mirrored in the tiffsep device.

With these changes in place Memento shows no more memory leaks with
this file and my guess at the configuration OSS-fuzz is using.