graphics library - fix memory leak with halftones
The handling of the 'order' structure from a device halftone, when
installing a halftone in the graphics state is complicated. There are
extensive comments at the head of gx_gstate_ht_install() but basically
we either 'take ownership' of the order and it's contents (specifically
the bit_data and levels arrays).

If we take ownership then we NULL the relevant portions of the structure
so that the caller doesn't free them.

If we don't take ownership then we make new copied of the data, and
leave the structure untouched, so that the caller cleans up.

However.... When deciding whether or not to take ownership we consider
whether the memory allocator used to create the data is the same as the
graphics state allocator. If it isn't then we must take a copy of the
data as we won't be able to free it when the halftone goes out of scope.

When the interpreter is pdfi the allocators don't match, so we copy the
data. But we then behave as if we had taken ownership by clearing the
structure, because we don't consider the different allocators. This
means pdfi doesn't clean up and we leak memory.

Fix that by checking mem_diff before clearing the structure members.
Note that, currently, used_default cannot be true if mem_diff is true so
the second check, at line 1152 is redundant, but it seems clearer to
keep it.