OSS-fuzz #49266 - avoid a divide-by-zero error
The TIFF devices Set the RowsPerStrip tag by dividing the max_strip_size
by the width of the raster. However it is entirely possible to set the
media size to a width of 0 pixels (when resolution is taken into
account), which causes a divide by zero.

Just check the width before dividing and if it is not at least 1 then
return an error.