GhostPDF - fix memory leak with Default* colour spaces
Exhibited with D:\bugs\Bugzilla_pdf\0692066_07366_1.pdf using either
nullpage or pkmraw.

We store the 'current' default colour spaces before we run a content
stream, and we put them back at the end, so that if the content stream
alters the Default* spaces we limit that change to the scope of the
content stream.

But, if it did change the Default* spaces while we correctly counted
down the 'new' spaces, we didn't count down the original ones when we
restored them. This caused the original spaces to leak.

So note that we changed spaces, and count down the old spaces after we
switch back. In effect giving up the references stored in the
PageDefault* local variables.