graphics library - fix memory leak with flattened paths
There are two conditions which mean we need to 'flatten' a path; firstly
if the path contains curves but also if any non-curve line is too long
to be handled in fixed point, in which case we break it into 2.

In either case we copy the original path and stroke the copy, then
discard it, leaving the original in place.

However, at the end of the routine we only considered whether the path
contained any curves when deciding whether we had copied the path (and
therefore need to free the copy). If we had a path which contained a
long line which had to be broken up, then we would copy the path, but
fail to free the copy leading to a memory leak.

The example file is the minimal test case attached to bug 692299 (a
MuPDF bug report). For me this exhibits a memory leak with pdfi and
the pkmraw device.