graphics library - Improve leak fix with flattened paths
It turns out (\tests\pdf\Bug6901014_Arioli-NAG-Warwick.pdf) that if a
path is flattened and then dashed the pointer 'spath' points to the
dashed path and no longer to the flattened path.

Stop trying to be clever and just set a flag if the path was flattened
and just use that to indicate that we need to free the flattened path.