GhostXPS - fix commit 7642e51271eafd012a80722b5f82226188de90a2
It seems that the logic was inverted here, we should check that the size
of the font is enough to satisfy the absolute minimum for the number
of fonts in the collection.

The previous commit was throwing an error if the length of the font was
more than the minimum. Basically all TrueType collections were detected
as being in error. This was causing incorrect output, non-deterministic
output and, on 32-bit tests, seg faults. The latter problems I find
surprising.....