GhostPDF - escape CR/LF in UTF16BE pdfmark strings for Acrobat
Bug #705686 "PDF Encryption makes bookmark garbled"

This appears to be a bug in Acrobat (but nothing to do with encryption).
The PDF file has an Outline which contains a Unicode (UTF-16BE) Title.
The first Unicode code point in the string is 0x54 0x0D, which the
PDF interpreter was passing through unchanged. The pdfwrite device
would then write the string without any further checking.

The result is that the string was written to the output as 0xFE 0xFF
0x54 0x0D. Which on the face of it is valid, it is permitted to have any
character in a string except the backslash '\' and unbalanced
parentheses '(' and ')'.

But when opened in Acrobat the glyph is displayed as if it were
0x54 0x0A. Change the string content to actually be 0x54 0x0A and it
continues to display the same wrong character. It seems like Acrobat
is detecting the 0x0D in the string and simply treating it as a 0x0A.

We can work around this by writing the 0x0D as an escaped character
instead '\r'. Acrobat happily draws the correct character. I chose to
add 0x0A as well, just in case.....