TIFF devices - restore default MaxStripSize
Commit 0abc209b8460396cdece8fc824c053a2662c4cbf altered the default
MaxStripSize from 8KB to 1MB, for all Tiff devices, even though it seems
that only fax readers have a problem with multiple strip TIFF files.

It also did not update the documentation which still stated the default
was 8KB.

After some discussion, it seems wrong that the default should have been
altered, when the documentation points to the recommendation in the TIFF
specification. Additionally the change was to a size of 1MB in the
'hope' that no fax page size would exceed that. Which is not a great way
to handle the problem.

This commit alters just the fax devices to have a default MaxStripSize
of 0, which means the output will be written to a single strip, no
matter how large, thus fulfilling the requirement to support poor fax
reading software, but without affecting the non-fax devices, and more
reliably because it doesn't matter how large the media is, even if it
exceeds 1MB output we will still only get a single strip..

If the reader can't cope with a single huge strip, and wants multiple
strips then the user can override the default by specifying
MaxStripSize on the command line in order to restore multiple-strip
output for fax devices.

Updated the documentation to reflect this.