GhostPDF - fix memory leak when dictionary boolean object is not a bool
file /bugzilla_pdf/0693651_09306_origen.pdf has a number of annotation
dictionaries where the entry should be a boolean false but is in fact a
name object '/False'.

When looking for a boolean to be retrieved from a dictionary, if the
returned object was neither a boolean true nor a boolean false we did
not count down the reference.