GhostPDF - cater for Encrypt entry being the null object
bugzilla_pdf\0687117_00346_pdf031031.pdf has an Encrypt entry in the
Root dictionary, but it is a null object which was causing us to error
trying to interpret it.

Just ignore null objects, but throw an error if the Encrypt entry is
neither a dictionary nor null.

Add an error entry, and one ofr the following commit to use too.