GhostPDF - better repairs for incorrect/invalid Root entries
File "d:\bugs\bugzilla_pdf\0688607_04666_foo2r.pdf has a Root dictionary
which does not contain the *required* /Type /Catalog key/value pair. It
does however contain the required /Pages key. So if we cannot find a
/Type key, don't immediately give up. If the dictionary contains a
/Pages key treat it as a Root dictionary and hope for the best.

File "d:\bugs\bugzilla_pdf\0689022_02651_eswc2005_spyns.pdf" has an
entry for the Root dictionary of "1 0 R", but this is completely wrong.
Object 1 is actually a Page dictionary and the Root dictionary is object
100. When repairing a file we do scan for the Catalog object and store
it in the context, so if we get a dictionary with a /Type key whose
value is not /Catalog, check to see if ctx->Root is non-NULL. If so we
have repaired the file and can try that dictionary. If not then we will
go round and repair the file, so we'll end up here again anyway.

File d:\bugs\bugzilla_pdf\0693719_09369_bug-poppler14303.pdf is a
similar problem, but in this case the Root entry in the trailer
dictionary points to an object which is not a dictionary. As above, if
we already have a ctx->Root, use that instead.

If we reach the end of repairing a file and code > 0 make it 0 as some
places check for == 0.

Not quite sure why, but these fixes exhibit significant progressions
with /tests_private/pdf/uploads/Bug696410.pdf