GhostPDF - attempt file repair if object number not in xref
If we get an object number larger than the number of objects in the
xref, instead of aborting with an error, rescan the file by running a
repair, in case it's the xref table which is broken.

File bugzilla_pdf\0698804_14526_xref-muckery.pdf has an xref table which
claims to have only 2 entries, but actually there are 5 objects in the
file. Repairing the file means we can open it.

This led to a seg fault with
/tests_private/pdf/sumatra/2064_-_mask_transfer_function_ignored.pdf
because we were not chcking the type of the SMask object before trying
to use it, so fix that too. This does lead to diffs with that file but
it is so badly corrupted that neither Acrobat nor the old Ghostscript
PDF interpreter will open it anyway.