oss-fuzz 49528: Free tiff context and related data correctly
Freeing the tiff "context" for a tiff file flushes the underlying file
descriptor, so we cannot close the file before doing so.

But freeing the tiff context also frees some memory that we need to access
in order to complete our part of closing the file.

So, we need to retrieve the file name we need, store it locally, free the
tiff file context, then close our file, in order to avoid accessing memory
after it's been freed.