OSS-fuzz 49541
If we fail to dereference an object we return the null object, but set
an error code. So we need to check the error, not just the fact that
fontfile is non-NULL because it might be the null object.

Note that this is a change because previously the null object was a real
object, and so would not have failed the dereference. It's now a small
integer and so does fail.