oss-fuzz 49530: Use allocated memory for custom enum procs
pdfwrite has to use a custom "set cache" proc when falling back to default
text handling to capture glyphs. It was using a locally declared structure
which obvious is problematic as the text enumerator outlasts the function.

So, allocate the procs structure dynanimically and add a custom "release" proc
to free the dynamic procs, as well as the enumerator itself.