GhostPDF - memory leak on error reading a stream
OSS-fuzz #42876 with Memento.

If we find a circular reference when checking the stream object number
we need to countdown the stream object before returning. We also should
clear the loop detector stack back to the mark we pushed. Finally we
need to clear the loop detector stack if we fail to add an object too.