OSS-fuzz #49606
If we rejected an addition to the xref table because the object number
exceeds a 32-bit signed integer, less than 0, or the generation number
is less than 0 or the file offset is less than 0 we were not returning
an error (earlier incarnations of the code didn't expect it to).

But when decoding an object stream if we didn't get an error code we
would assume the object was present in the xref table and try to use
the xref  entry.

Fix that by returning an error code if we didn't add an entry to the
xref table.