Windows build - Fix versioning of the binaries
It seems that the same information is stored in two ways in the version
resource of Windows binaries. As numeric data and as a human-readable
string.

With the addition of the patch level we had updated the string data but
had omitted to update the numeric information. In addition it appears
that we always had the word order reversed, so that instead of (eg)
9.57.1.0 our numeric version was coming out as 0.0.9.57 (wrong order
and no patch level, if the patch level had been present this would
have read 1.0.9.57).

This commit corrects the word order of FILEVERSION and PRODUCTVERSION
and adds the patch level to them, which means we also need to define
the patch level on the invocation of the Resource Compiler.