Bug #705699
" “Error: /invalidfileaccess in --file--” when converting PBM file to PDF"

There is no actual bug here, the user had not supplied either -dNOSAFER
or --permit-file-read on the command line to allow the interpreter to
read the input file.

However the viewpbm.ps program doesn't have that information in its
'usage' which would be useful information. So added that here, and
updated a number of other view*.ps programs which have usage information
to include this as well.