pdfwrite - don't try to use a NULL pointer as a cos_dict
OSS-fuzz #49694

The horrible cos_dict implementation doesn't check to see if the pointer
it has received is valid before dereferencing it, it just assumes it
will be valid.

In this case it is NULL (the input file is terribly broken), and trying
to use cos_dict_find on a NULL pointer seg faults quickly when it tries
to dereference it.

In this commit I've modified all the places in the failing routine which
try to use the NULL pointer so that they test it first. I've also
modified cos_dict_find so that it returns NULL instead of seg faulting
in this case. Sadly there are many other places which look vulnerable
in the cos_dict mess. Ideally I'd review those now and fix them all but
I don't really have the time. I'll try and review it when I have some
time spare.