GhostPDF - yet more circular references in Info dictionaries
OSS-fuzz #49703

OSS-fuzz finds another way to evade the circular reference checking on
the Info dictionary, this time by having Annots point to the Pages
tree instead of the Page, **and** having the Info dictionary point to
the first Page dictionary.

I was going to dump anything except names and strings from the
dictionary, forgetting that we use the same mechanism for returning the
Page dictionary, and we need to have that able to deal with arrays at
least, for the various Boxes.

So add yet more checking to the contents of the PDF object we are being
asked to turn into a PostScript object.