Bug 705696: Issues with -sICCOutputColors and Transparency
The pdf14 device was avoiding carrying around extra
bands that are not part of the source spot colorants, but this
becomes a bookkeeping headache that is a source of too many bugs,
is difficult to maintain, and prone to problems. To avoid these problems,
the pdf14 device will now maintain all the spot colorants of the target
device.  In some respects this may be useful if someone does want to
use these "extra" spot planes for use withing the pdf14 compositor,
they will have the opportunity with this change.  Note that Bug 7095696
also has a bug related to the psdcmyktags devices.  That is not fixed
with this commit, but will be dealt with in a separate commit.