Bug 705696 psdcmyktags (and psdcmyktags16) crash
The spot color sorting code in the psdcmyk device was not
correctly accounting for the tag plane in its process. With
this fix we no longer get a crash. However, it is noted that
for files with spot colors the psdcmyktags device is not
rendering correctly.  I am going to open a separate bug for this
and close 705696 since it already has two separate issues in it.