GhostPDF - fix memory leak with type 1 /Subrs
We were counting up the newly created string (which is correct because
new object have a ref count of 0) then storing it in the array, which
increments its reference count to 2. But we did not count down the
reference to the string which meant that when we freed the array we did
not free the string, and it leaked.