Graphics library - fix heap-buffer overflow in 8-bit memory device
Bug #705705 "heap-buffer-overflow in base/gdevm8.c:222 in mapped8_copyN1"

The loop copying from source to destination updates a local variable
from the source pointer at the end of the loop rather than at the start.
This is because although the input is single bits, the loop only
operates on whole bytes, so after the loop completes there may be more
bits to process, which are stored in the top part of the next byte.

The problem is that if the raster happens to fit exactly into a byte
then when we finish the final byte we read the byte one beyond the end
of the buffer. We don't actually use that byte for anything but we do
read it.

This commit refactors the code so that we read the byte at the beginning
of the loop (but before any of the labels, so that the goto above the
loop will work properly). Then if there are any remaining bits after
the loop terminates we read an extra byte.

This does introduce one extra check against 0 per scan line, but I can't
see an easy way to avoid that without rewriting the whole function.