graphics library - another heap-buffer overflow in copy_mono
Bug #705707 "heap-buffer-overflow in gdevp14.c:6345 pdf14_copy_mono"

As with the previous case, if the number of bits happens to be an exact
multiple then we read a byte in preparation for the next loop, but
never use it. This in itself is a bad thing.

So here we once again move the byte reading from the end of the loop to
the beginning so that we don't read the next byte if we aren't actually
going to use it. To ensure we don't skip past the first byte, and
because the first thing we will do is decrement 'bit', start by adding
1 to it so it all evens out.