GhostPDF - fix recursion with ObjStm
Bug #705711 "stack overflow in pdf_deref.c:845 pdfi_dereference_main (exploitable)"

This has an ObjStm (compressed object stream) which has a /Length which
is an indirect object. When we consult the xref we find that object is
itself contained in an ObjStm. In fact, the same ObjStm.....

This commit reorders some of the dictionary accesses so they all occur
in one place, and adds loop detection to the dereferencing of those
objects, if loop detection is going on (it almost always is), and adds
the ObjStm to the list of objects being tracked for the duration of
reading it's own dictionary. So if it is indirectly self-referencing
then we will now detect it as a circular reference and throw an error.