Bug 705717: Gray settings in -sSourceObjectICC
The gray profiles were not getting set properly when we set the option
to NONE in the SourceObjectICC file for Text, Graphics and Images.

Note that the interaction of -sSourceObjectICC and -dDeviceGrayToK=true/false
is as follows:

If no gray settings are made in SourceObjectICC, and there is DeviceGray
content present in the source file, -dDeviceGrayToK=true/false will be honored.
If the gray objects in SourceObjectICC are set to None, which implies that
ICC color management is not to be applied to these objects, then they
are treated as DeviceGray and always mapped to K values in a CMYK target device,
regardless of the settings of -dDeviceGrayToK (i.e. there is no color
management).

If instead, the gray objects in SourceObjectICC are set to a specific ICC
profile, then they are no longer DeviceGray but are ICC colors. They will
be color managed, regardless of the setting of -dDeviceGrayToK.