GhostPDF - permit array argument for Blend Mode
Bug #705718 "Blend Modes with array arguments not working"

The file has an ExtGState with a /BM which is an array not a name.
Astonishingly this is valid and we are supposed to use the first entry
in the array (of names) which we recognise, or /Normal if we don't
recognise any (or if none of them are names, presumably).