GhostPDF - when validating Info, reset the dictionary on key deletion
Bug #705710 "stack-overflow in idict.c:491 dict_unpack"

When we found an invalid (circular reference) key/value pair in the
Info dictionary we deleted the pair and carried on. But we can't carry
on iterating through the dictionary after we've deleted a pair and that
caused an error, which meant that subsequent invalid pairs were not
sanitized.

Instead, after we delete a key/value pair, reset to the beginning and
rescan the dictionary. This will be slower, but only for dictionaries
which contain circular references and they are broken, so....