Bug 705733: Improve pattern error cleanup
Type 1 patterns are loaded lazily, and normal cleanup occurs when the pattern
tile is ejected from the tile cache.

But in this case, the pattern triggers an error in the transparency code,
meaning the pattern tile is never added to the cache. This causes a memory leak
in gpdf, and a crash in the gs garbage collector when shutting down the PS
interpreter (when pdfi is called via gs).

We now ensure the pattern accumulator device is destroyed in the event of such
an error, meaning no spurious references to the pattern, the target device nor
the contents of the pattern instance remain to cause problems.